chore: refactor migration location
