+++ drizzle.config.ts
@@ -1,13 +1,13 @@
-import { type Config } from "drizzle-kit";
+import { type Config } from 'drizzle-kit'
 
-import { env } from "~/env";
+import { env } from '~/env'
 
 export default {
-  schema: "./src/migration/schema.ts",
-  driver: "pg",
+  schema: './src/server/db/migration/schema.ts',
+  driver: 'pg',
   dbCredentials: {
     connectionString: env.DATABASE_URL,
   },
   // tablesFilter: ["chatbot-builder_*"],
-  out: "./migrations",
-} satisfies Config;
+  out: './src/server/db/migrations',
+} satisfies Config
